requirements.txt
llama-index
deeplake
python-dotenv
llama_hub
streamlit
deeplake
replicate